+++ components/collector/tests/source_collectors/testng/test_tests.py
@@ -26,20 +26,25 @@ class TestNGTestsTest(TestNGCollectorTestCase):
       </suite>
     </testng-results>
     """
+    EXPECTED_ENTITIES = [
+        dict(key="class1_method1", class_name="class1", name="method1", test_result="passed"),
+        dict(key="class1_method2", class_name="class1", name="method2", test_result="passed"),
+        dict(key="class1_method3", class_name="class1", name="method3", test_result="failed"),
+    ]
 
     async def test_tests(self):
         """Test that the number of tests is returned."""
         response = await self.collect(get_request_text=self.TESTNG_XML)
-        self.assert_measurement(response, value="3", total="3")
+        self.assert_measurement(response, value="3", total="3", entities=self.EXPECTED_ENTITIES)
 
     async def test_failed_tests(self):
         """Test that the failed tests are returned."""
         self.set_source_parameter("test_result", ["failed"])
         response = await self.collect(get_request_text=self.TESTNG_XML)
-        self.assert_measurement(response, value="1", total="3")
+        self.assert_measurement(response, value="1", total="3", entities=self.EXPECTED_ENTITIES[-1:])
 
     async def test_zipped_testng_report(self):
         """Test that the number of tests is returned from a zip with TestNG reports."""
         self.set_source_parameter("url", "testng.zip")
         response = await self.collect(get_request_content=self.zipped_report(("testng.xml", self.TESTNG_XML)))
-        self.assert_measurement(response, value="3", total="3")
+        self.assert_measurement(response, value="3", total="3", entities=self.EXPECTED_ENTITIES)